app/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
importlib-resources==5.4.0
mysqlclient==2.1.1
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.1.1
utils==1.0.2
zipp==3.6.0
